+++ example/lib/LayerGraphView.dart
@@ -52,7 +52,7 @@ class _LayeredGraphViewPageState extends State<LayeredGraphViewPage> {
             ),
             RaisedButton(
               onPressed: () {
-                final Node node12 = Node(getNodeText());
+                final node12 = Node(rectangleWidget(r.nextInt(100)));
                 var edge = graph.getNodeAtPosition(r.nextInt(graph.nodeCount()));
                 print(edge);
                 graph.addEdge(edge, node12);
@@ -75,6 +75,11 @@ class _LayeredGraphViewPageState extends State<LayeredGraphViewPage> {
                   ..color = Colors.green
                   ..strokeWidth = 1
                   ..style = PaintingStyle.stroke,
+                builder: (Node node) {
+                  // I can decide what widget should be shown here based on the id
+                  var a = node.key.value as int;
+                  return rectangleWidget(a);
+                },
               )),
         ),
       ],
@@ -83,9 +88,7 @@ class _LayeredGraphViewPageState extends State<LayeredGraphViewPage> {
 
   Random r = Random();
 
-  int n = 1;
-
-  Widget getNodeText() {
+  Widget rectangleWidget(int a) {
     return Container(
         padding: EdgeInsets.all(16),
         decoration: BoxDecoration(
@@ -94,7 +97,7 @@ class _LayeredGraphViewPageState extends State<LayeredGraphViewPage> {
             BoxShadow(color: Colors.blue[100], spreadRadius: 1),
           ],
         ),
-        child: Text("Node ${n++}"));
+        child: Text("Node ${a}"));
   }
 
   final Graph graph = Graph();
@@ -103,29 +106,29 @@ class _LayeredGraphViewPageState extends State<LayeredGraphViewPage> {
 
   @override
   void initState() {
-    final Node node1 = Node(getNodeText());
-    final Node node2 = Node(getNodeText());
-    final Node node3 = Node(getNodeText());
-    final Node node4 = Node(getNodeText());
-    final Node node5 = Node(getNodeText());
-    final Node node6 = Node(getNodeText());
-    final Node node8 = Node(getNodeText());
-    final Node node7 = Node(getNodeText());
-    final Node node9 = Node(getNodeText());
-    final Node node10 = Node(getNodeText());
-    final Node node11 = Node(getNodeText());
-    final Node node12 = Node(getNodeText());
-    final Node node13 = Node(getNodeText());
-    final Node node14 = Node(getNodeText());
-    final Node node15 = Node(getNodeText());
-    final Node node16 = Node(getNodeText());
-    final Node node17 = Node(getNodeText());
-    final Node node18 = Node(getNodeText());
-    final Node node19 = Node(getNodeText());
-    final Node node20 = Node(getNodeText());
-    final Node node21 = Node(getNodeText());
-    final Node node22 = Node(getNodeText());
-    final Node node23 = Node(getNodeText());
+    final node1 = Node.Id(1);
+    final node2 = Node.Id(2);
+    final node3 = Node.Id(3);
+    final node4 = Node.Id(4);
+    final node5 = Node.Id(5);
+    final node6 = Node.Id(6);
+    final node8 = Node.Id(7);
+    final node7 = Node.Id(8);
+    final node9 = Node.Id(9);
+    final node10 = Node.Id(10);
+    final node11 = Node.Id(11);
+    final node12 = Node.Id(12);
+    final node13 = Node.Id(13);
+    final node14 = Node.Id(14);
+    final node15 = Node.Id(15);
+    final node16 = Node.Id(16);
+    final node17 = Node(rectangleWidget(17)); //using deprecated mechanism of directly placing the widget here
+    final node18 = Node(rectangleWidget(18));
+    final node19 = Node(rectangleWidget(19));
+    final node20 = Node(rectangleWidget(20));
+    final node21 = Node(rectangleWidget(21));
+    final node22 = Node(rectangleWidget(22));
+    final node23 = Node(rectangleWidget(23));
 
     graph.addEdge(node1, node13, paint: Paint()..color = Colors.red);
     graph.addEdge(node1, node21);